% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{aggregate.Matrix}
\alias{aggregate.Matrix}
\title{aggregate.Matrix
for each column estimate per group (of rows) mean}
\usage{
\method{aggregate}{Matrix}(x, group, fun = "sum")
}
\arguments{
\item{x}{a \code{n} x \code{p} Matrix or object coercible to one}

\item{group}{a factor or a object coercible to one specifying the grouping of rows of \code{x} into \code{k} groups}
}
\value{
a \code{k} x \code{p} [\link[Matrix:dgCMatrix]{Matrix::dgCMatrix}] of values of \code{x} aggragated by \code{group}
}
\description{
aggregate.Matrix
for each column estimate per group (of rows) mean
}
