% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{aggregate_var.Matrix}
\alias{aggregate_var.Matrix}
\title{aggregate_var.Matrix
for each column estimate per group (of rows) variance}
\usage{
aggregate_var.Matrix(x, group)
}
\arguments{
\item{x}{a Matrix or object coercible to one}

\item{group}{a factor or a object coercible to one specifying the grouping of rows of \code{x}}
}
\description{
aggregate_var.Matrix
for each column estimate per group (of rows) variance
}
\seealso{
Matrix.utils::aggregate.Matrix
}
