% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixcolorr.R
\name{col2sRGB}
\alias{col2sRGB}
\title{Convert "R" colors to colorspaces' sRGB colors}
\usage{
col2sRGB(col)
}
\arguments{
\item{col}{A character vector of color names or hexadecimal color codes.}
}
\value{
An object of class \link[colorspace:sRGB]{"sRGB"} containing the red, green, and blue components of the color.
}
\description{
This function takes R color strings and converts them into colorspaces::sRGB values.
It utilizes grDevices::col2rgb for conversion.
}
\examples{
col2sRGB("red")
col2sRGB("#FF0000")
}
