% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{deming}
\alias{deming}
\title{Deming regression for ggplot stat_smooth and geom_smooth
With bootstrapped confidence intervals}
\usage{
deming(formula, data, R = 100, noise_ratio = NULL, ...)
}
\arguments{
\item{formula}{formula for the tls}

\item{data}{Input data.frame}

\item{noise_ratio}{numeric scalar of the measurement error of LHS over that of RHS if missing or NULL}

\item{...}{Unused argument}
}
\description{
Deming regression for ggplot stat_smooth and geom_smooth
With bootstrapped confidence intervals
}
