% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_split_tile.R
\name{draw_key_split_tile}
\alias{draw_key_split_tile}
\title{Provides a diagonally split version of GeomTile}
\usage{
draw_key_split_tile(data, params, size)
}
\description{
based on ggplot2::GeomTile by the ggplot2 authors https://github.com/tidyverse/ggplot2/graphs/contributors
}
