% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_scales.R
\name{log_eng}
\alias{log_eng}
\alias{scale_x_log_eng}
\alias{scale_x_log}
\alias{scale_y_log_eng}
\alias{scale_y_log}
\alias{log1p_breaks}
\alias{scale_x_log1p}
\alias{scale_y_log1p}
\title{Nice log / log1p Scale
with radix dependent breaks.}
\usage{
scale_x_log_eng(..., radix = 10)

scale_x_log(..., radix = 10)

scale_y_log_eng(..., radix = 10)

scale_y_log(..., radix = 10)

log1p_breaks(maj, radix = 10)

scale_x_log1p(..., radix = 10)

scale_y_log1p(..., radix = 10)
}
\description{
Nice log / log1p Scale
with radix dependent breaks.
}
