% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{long2sparse}
\alias{long2sparse}
\title{Create sparse matrix from long data.table}
\usage{
long2sparse(
  rows,
  cols,
  values,
  dimname_rows = base::deparse1(substitute(rows)),
  dimname_cols = base::deparse1(substitute(rows))
)
}
\description{
Create sparse matrix from long data.table
}
\examples{
long2sparse(1:3, 2:4, runif(3))
}
