% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixcolorr.R
\name{mixcolorr}
\alias{mixcolorr}
\title{Mix two "R" colors}
\usage{
mixcolorr(alpha, col1, col2, ...)
}
\arguments{
\item{alpha}{A numeric vector of length 1 or n with values between 0 (only col1) and 1 (only col2) indicating the mixing coefficient.}

\item{col1}{The first color to mix, as a name or hexadecimal code.}

\item{col2}{The second color to mix, as a name or hexadecimal code.}

\item{...}{other arguments passed to \code{\link[colorspace:mixcolor]{colorspace::mixcolor()}}}
}
\value{
A hexadecimal color code representing the mixed color.
}
\description{
\code{mixcolorr} mixes two R color strings in a given proportion. It is a wrapper
around \code{\link[colorspace:mixcolor]{colorspace::mixcolor()}} that accepts and returns \link[grDevices:col2rgb]{R color strings} like
\code{"red"} or\code{"#FF0000"}. It is vectorized.
}
\examples{
mixcolorr(0.8, "red", "gray")
}
\seealso{
\code{\link[colorspace:mixcolor]{colorspace::mixcolor()}} \code{\link[grDevices:colors]{grDevices::colors()}} \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}}
}
