% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{predictdf.Deming}
\alias{predictdf.Deming}
\title{prediction function for ggplot2's stat_smooth}
\usage{
predictdf.Deming(model, xseq, se, level)
}
\arguments{
\item{model}{Input model}

\item{xseq}{x-values used for prediction}

\item{se}{Predict error or not}

\item{level}{Confidence level}
}
\description{
prediction function for ggplot2's stat_smooth
}
