% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helpers.R
\name{quantile_cl}
\alias{quantile_cl}
\title{compute confidence intervals for a quantile (e.g. median)
See also https://stats.stackexchange.com/questions/122001/confidence-intervals-for-median}
\usage{
quantile_cl(y, q = 0.5, conf.level = 0.95, na.rm = TRUE)
}
\description{
compute confidence intervals for a quantile (e.g. median)
See also https://stats.stackexchange.com/questions/122001/confidence-intervals-for-median
}
