% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helpers.R
\name{scale_fill_symetric}
\alias{scale_fill_symetric}
\title{Continuous fill scale that is symmetric around 0}
\usage{
scale_fill_symetric(
  ...,
  limit = function(range) max(abs(range)),
  limits = function(range) (if (is.function(limit)) limit(range) else range) * c(-1, 1)
)
}
\arguments{
\item{...}{arguments passed to  \code{\link[ggplot2:scale_colour_continuous]{ggplot2::scale_fill_continuous()}}}

\item{limit}{Either a function of the data range or a scalar, defining the upper limit used to compute the limits. ignored if \code{limits} is specified.}

\item{limits}{the limits passed to \code{\link[ggplot2:scale_colour_continuous]{ggplot2::scale_fill_continuous()}}}
}
\description{
Continuous fill scale that is symmetric around 0
}
\seealso{
\code{\link[ggplot2:scale_colour_continuous]{ggplot2::scale_fill_continuous()}}
}
