% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schelpr-package.R
\docType{package}
\name{schelpr-package}
\alias{schelpr-package}
\title{schelpr: scHelpR - single cell helper package}
\description{
A random collection of helper functions that have yet to make it into appropriate packages. Use with care, consider copying individual functions to you codebase
}
\author{
\strong{Maintainer}: Jan Gleixner \email{jan.gleinxer@gmail.com} (\href{https://orcid.org/0000-0001-5006-7375}{ORCID})

}
\keyword{internal}
