% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schelpr.R
\docType{package}
\name{schelpr}
\alias{schelpr}
\title{Helper package for analysis of CROP-seq-ULI data}
\description{
pronounced "shellp R". Contains a random collection of helper functions that have yet to make it into appropriate packages
}
\section{plotting}{
NA
}

\section{Seurat}{
NA
}

\section{other}{
NA
}

