% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sparse2long}
\alias{sparse2long}
\title{Create data.table from sparse matrix}
\usage{
sparse2long(mat, value_name = "value")
}
\description{
Create data.table from sparse matrix
}
\examples{
sparse2long(Matrix::spMatrix(1:3,2:4,3:5, ncol=4, nrow=3))
}
