% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helpers.R
\name{stat_bin_2d_integer}
\alias{stat_bin_2d_integer}
\title{stat_bin_2d_integer}
\usage{
stat_bin_2d_integer(
  mapping = NULL,
  data = NULL,
  geom = "rect",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  nbins = 30,
  ...
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with aes or aes_.}

\item{data}{The data to be displayed in the layer.}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them.}

\item{...}{other arguments passed on to layer.}

\item{bins}{number of bins}
}
\value{
ggplot layer
}
\description{
A function to use the StatBin2dInteger statistic
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, disp)) +
  stat_bin_2d_integer(aes(fill = after_stat(count)), nbins = 30) +
  scale_y_log10()
}
