% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helpers.R
\docType{data}
\name{StatBin2dInteger}
\alias{StatBin2dInteger}
\title{StatBin2dInteger}
\format{
An object of class \code{StatBin2dInteger} (inherits from \code{StatBin2d}, \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatBin2dInteger
}
\description{
A 2d binning statistic that calculates bins by rounding to integer values
}
\keyword{datasets}
