% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_scales.R
\name{format_selected_breaks}
\alias{format_selected_breaks}
\alias{format_major_breaks}
\alias{log_breaks}
\alias{logX_trans}
\alias{scale_x_log}
\alias{scale_y_log}
\alias{scale_color_log}
\alias{scale_fill_log}
\alias{scale_x_log_eng}
\alias{scale_y_log_eng}
\alias{log1p_scaled_breaks}
\alias{log1p_scaled_trans}
\alias{scale_x_log1p_scaled}
\alias{scale_y_log1p_scaled}
\alias{scale_color_log1p_scaled}
\alias{scale_fill_log1p_scaled}
\alias{log1p_breaks}
\alias{log1p_trans}
\alias{scale_x_log1p}
\alias{scale_y_log1p}
\alias{scale_color_log1p}
\alias{scale_fill_log1p}
\title{Nice log / log1p / log1pscaled scales
with radix dependent breaks.}
\usage{
format_selected_breaks(breaks_to_label, format)

format_major_breaks(breaks_to_label, format)

log_breaks(maj, radix = 10, max_major_breaks_to_show_minor = 4)

logX_trans(
  radix,
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE),
  format = scales::format_format()
)

scale_x_log(
  ...,
  radix = 10,
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)

scale_y_log(
  ...,
  radix = 10,
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)

scale_color_log(
  ...,
  radix = 10,
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)

scale_fill_log(
  ...,
  radix = 10,
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)

scale_x_log_eng(
  ...,
  radix = 10,
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)

scale_y_log_eng(
  ...,
  radix = 10,
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)

log1p_scaled_breaks(maj, radix = 10, scale = 1)

log1p_scaled_trans(
  radix,
  scale,
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE),
  format = scales::format_format()
)

scale_x_log1p_scaled(
  ...,
  radix = 10,
  scale = 1,
  expand = expansion(mult = c(0, 0.05)),
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)

scale_y_log1p_scaled(
  ...,
  radix = 10,
  scale = 1,
  expand = expansion(mult = c(0, 0.05)),
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)

scale_color_log1p_scaled(
  ...,
  radix = 10,
  scale = 1,
  expand = expansion(mult = c(0, 0.05)),
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)

scale_fill_log1p_scaled(
  ...,
  radix = 10,
  scale = 1,
  expand = expansion(mult = c(0, 0.05)),
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)

log1p_breaks(maj, radix = 10)

log1p_trans(
  radix = 10,
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE),
  format = scales::format_format()
)

scale_x_log1p(
  ...,
  radix = 10,
  expand = expansion(mult = c(0, 0.05)),
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)

scale_y_log1p(
  ...,
  radix = 10,
  expand = expansion(mult = c(0, 0.05)),
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)

scale_color_log1p(
  ...,
  radix = 10,
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)

scale_fill_log1p(
  ...,
  radix = 10,
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)

scale_color_log1p(
  ...,
  radix = 10,
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)

scale_fill_log1p(
  ...,
  radix = 10,
  show_minor_ticks = getOption("shelpr.show_minor_ticks", TRUE)
)
}
\description{
Nice log / log1p / log1pscaled scales
with radix dependent breaks.
}
