% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_scales.R
\name{log_breaks}
\alias{log_breaks}
\alias{log_eng}
\alias{scale_x_log_eng}
\alias{scale_x_log}
\alias{scale_y_log_eng}
\alias{scale_y_log}
\alias{log1p_scaled_breaks}
\alias{format_major_breaks}
\alias{log1p_scaled_trans}
\alias{scale_x_log1p_scaled}
\alias{scale_y_log1p_scaled}
\alias{scale_color_log1p_scaled}
\alias{log1p_breaks}
\alias{log1p_trans}
\alias{scale_x_log1p}
\alias{scale_y_log1p}
\title{Nice log / log1p Scale
with radix dependent breaks.}
\usage{
log_breaks(maj, radix = 10)

scale_x_log_eng(
  ...,
  breaks = log_breaks(TRUE, radix),
  minor_breaks = log_breaks(FALSE, radix),
  radix = 10
)

scale_x_log(
  ...,
  breaks = log_breaks(TRUE, radix),
  minor_breaks = log_breaks(FALSE, radix),
  radix = 10
)

scale_y_log_eng(
  ...,
  breaks = log_breaks(TRUE, radix),
  minor_breaks = log_breaks(FALSE, radix),
  radix = 10
)

scale_y_log(
  ...,
  breaks = log_breaks(TRUE, radix),
  minor_breaks = log_breaks(FALSE, radix),
  radix = 10
)

log1p_scaled_breaks(maj, radix = 10, scale = 1)

format_major_breaks(breaks_to_label, format)

log1p_scaled_trans(radix, scale, format = scales::format_format())

scale_x_log1p_scaled(..., radix = 10, scale = 1)

scale_y_log1p_scaled(..., radix = 10, scale = 1)

scale_color_log1p_scaled(..., radix = 10, scale = 1)

log1p_breaks(maj, radix = 10)

log1p_trans(radix, format = scales::format_format())

scale_x_log1p(..., radix = 10)

scale_y_log1p(..., radix = 10)
}
\description{
Nice log / log1p Scale
with radix dependent breaks.
}
